+++ GoedeDoelenHelpen/ReactApp/src/App.tsx
@@ -1,86 +1,16 @@
-/*import * as React from 'react';
-import './App.css';
-
-const logo = require('./logo.svg');
-
-class App extends React.Component {
-  render() {
-    return (
-      <div className="App">
-        <header className="App-header">
-          <img src={logo} className="App-logo" alt="logo" />
-          <h1 className="App-title">Welcome to React</h1>
-        </header>
-        <p className="App-intro">
-          To get started, edit <code>src/App.tsx</code> and save to reload.
-        </p>
-      </div>
-    );
-  }
-}
-
-export default App;*/
-
 import * as React from 'react';
 import {
   BrowserRouter as Router,
   Route,
   Link,
-  RouteComponentProps
-  
+  Switch
 } from 'react-router-dom';
+import about from './pages/about';
+import Home from './pages/Home';
+import Topics from './pages/Topics';
+import Page404 from './pages/Page404';
 
-const Home = () => (
-  <div>
-    <h2>Home</h2>
-  </div>
-);
-
-const About = () => (
-  <div>
-    <h2>About</h2>
-  </div>
-);
-
-const Topic = ({ match }: RouteComponentProps<{topicId: string}>) => (
-  <div>
-    <h3>{match.params.topicId}</h3>
-  </div>
-);
-
-const Topics = ({ match }: RouteComponentProps<{}>) => (
-  <div>
-    <h2>Topics</h2>
-    <ul>
-      <li>
-        <Link to={`${match.url}/rendering`}>
-          Rendering with React
-        </Link>
-      </li>
-      <li>
-        <Link to={`${match.url}/components`}>
-          Components
-        </Link>
-      </li>
-      <li>
-        <Link to={`${match.url}/props-v-state`}>
-          Props v. State
-        </Link>
-      </li>
-    </ul>
-
-    <Route path={`${match.path}/:topicId`} component={Topic}/>
-    <Route 
-      exact={true}
-      path={match.path} 
-      render={() => (
-      <h3>Please select a topic.</h3>
-    )}
-    />
-  </div>
-);
-
-const BasicExample = () => (
+const App = () => (
   <Router>
     <div>
       <ul>
@@ -91,10 +21,13 @@ const BasicExample = () => (
 
       <hr/>
 
-      <Route exact={true} path="/" component={Home}/>
-      <Route path="/about" component={About}/>
-      <Route path="/topics" component={Topics}/>
+      <Switch>
+        <Route exact={true} path="/" component={Home}/>
+        <Route path="/about" component={about}/>
+        <Route path="/topics" component={Topics}/>
+        <Route component={Page404} />
+      </Switch>
     </div>
   </Router>
 );
-export default BasicExample;
+export default App;
