+++ GoedeDoelenHelpen/ReactApp/src/pages/Home.tsx
@@ -1,0 +1,7 @@
+import * as React from 'react';
+
+export default () => (
+    <div>
+      <h2>Home</h2>
+    </div>
+  );
