+++ GoedeDoelenHelpen/ReactApp/src/pages/Page404.tsx
@@ -1,0 +1,7 @@
+import * as React from 'react';
+
+export default () => (
+    <div>
+        <p>Page not found</p>
+    </div>
+    );
