+++ GoedeDoelenHelpen/ReactApp/src/pages/about.tsx
@@ -1,0 +1,7 @@
+import * as React from 'react';
+
+export default () => (
+    <div>
+      <h2>About</h2>
+    </div>
+  );
